requirements-neuron.txt
# Common dependencies
-r requirements-common.txt

# Dependencies for Neuron devices
transformers-neuronx >= 0.9.0
torch-neuronx >= 2.1.0
neuronx-cc